# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/auth/refresh": {
            "post": {
                "description": "使用Refresh Token获取新的Access Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "刷新Access Token",
                "parameters": [
                    {
                        "description": "Refresh Token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刷新成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "无效的Refresh Token",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "backend_pkg_response.ErrorField": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "backend_pkg_response.PageInfo": {
            "type": "object",
            "properties": {
                "page_num": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "backend_pkg_response.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "error": {
                    "$ref": "#/definitions/backend_pkg_response.ErrorField"
                },
                "page": {
                    "$ref": "#/definitions/backend_pkg_response.PageInfo"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
    }
}